% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{identity}
\alias{identity}
\title{Return an argument}
\usage{
identity(x)
}
\arguments{
\item{x}{The only argument.}
}
\value{
\code{x} unmodified.
}
\description{
This is the only function in the \code{CLPFUDatabase}
metapackage.
Its purpose is to avoid warnings
during the R CMD check process.
}
\examples{
identity(42)
}
